Moving store service into store folder
